requirements.txt
requests
python-dotenv
pydub